Remove boxing requirement for Mojo response callbacks.
Now that the callbacks no longer inherit from generic types, reply
parameters can directly use primitive types for response parameters. A
few response callbacks have also been converted to use the preferred
lambda syntax where it makes sense.

Bug: 1491717, 1491924,
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5142600
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Owners-Override: Oksana Zhuravlova <oksamyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240352}

+++ components/webauthn/android/java/src/org/chromium/components/webauthn/InternalAuthenticator.java
@@ -108,7 +108,7 @@ public class InternalAuthenticator {
                         InternalAuthenticatorJni.get()
                                 .invokeMakeCredentialResponse(
                                         mNativeInternalAuthenticatorAndroid,
-                                        status.intValue(),
+                                        status,
                                         response == null ? null : response.serialize());
                     }
                 });
@@ -131,7 +131,7 @@ public class InternalAuthenticator {
                         InternalAuthenticatorJni.get()
                                 .invokeGetAssertionResponse(
                                         mNativeInternalAuthenticatorAndroid,
-                                        status.intValue(),
+                                        status,
                                         response == null ? null : response.serialize());
                     }
                 });
